Move Assertions from util to internal.
