+++ rxandroid/src/main/java/rx/android/internal/Assertions.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package rx.android.util;
+package rx.android.internal;
 
 import android.os.Looper;
