+++ rxandroid/src/main/java/rx/android/view/OperatorViewClick.java
@@ -17,7 +17,7 @@ import android.view.View;
 import rx.Observable;
 import rx.Subscriber;
 import rx.Subscription;
-import rx.android.util.Assertions;
+import rx.android.internal.Assertions;
 import rx.android.AndroidSubscriptions;
 import rx.functions.Action0;
