+++ rxandroid/src/main/java/rx/android/widget/OnSubscribeListViewScroll.java
@@ -18,7 +18,7 @@ import android.widget.AbsListView;
 import android.widget.AdapterView;
 import rx.Observable;
 import rx.Subscriber;
-import rx.android.util.Assertions;
+import rx.android.internal.Assertions;
 import rx.android.AndroidSubscriptions;
 import rx.functions.Action0;
