+++ rxandroid/src/main/java/rx/android/widget/OperatorCompoundButtonInput.java
@@ -19,7 +19,7 @@ import rx.Observable;
 import rx.Subscriber;
 import rx.Subscription;
 import rx.android.view.OnCheckedChangeEvent;
-import rx.android.util.Assertions;
+import rx.android.internal.Assertions;
 import rx.android.AndroidSubscriptions;
 import rx.functions.Action0;
