+++ rxandroid/src/main/java/rx/android/widget/OperatorTextViewInput.java
@@ -16,7 +16,7 @@ package rx.android.widget;
 import rx.Observable;
 import rx.Subscriber;
 import rx.Subscription;
-import rx.android.util.Assertions;
+import rx.android.internal.Assertions;
 import rx.android.AndroidSubscriptions;
 import rx.functions.Action0;
 import android.text.Editable;
